Refactor: 상수 모듈 consts 폴더로 이동 및 join 관련 메서드 수정
- CRUD.js : 타입가드 제거
- Parser.js : from 메서드 대신 wrapBacktick 사용
- Validator.js : Error 모듈 이동으로 인한 import 문 수정
- Join.js : Join 상수 타입 모듈
- Error.js : join.type 에러 메시지 수정
