+++ src/consts/Join.js
@@ -1,0 +1,4 @@
+export const JOIN = {
+    LEFT: "LEFT",
+    RIGHT: "RIGHT",
+};
